Fixes issue #2422 - Add Jacoco check to TEMPDIR extension
