--- extension/standard-tempdir/src/main/java/cloud/piranha/extension/standard/tempdir/StandardTempDirExtension.java
@@ -1,72 +1,0 @@
-/*
- * Copyright (c) 2002-2022 Manorrock.com. All Rights Reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- *   1. Redistributions of source code must retain the above copyright notice,
- *      this list of conditions and the following disclaimer.
- *   2. Redistributions in binary form must reproduce the above copyright
- *      notice, this list of conditions and the following disclaimer in the
- *      documentation and/or other materials provided with the distribution.
- *   3. Neither the name of the copyright holder nor the names of its
- *      contributors may be used to endorse or promote products derived from
- *      this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- */
-package cloud.piranha.extension.standard.tempdir;
-
-import cloud.piranha.core.api.WebApplication;
-import cloud.piranha.core.api.WebApplicationExtension;
-import jakarta.servlet.ServletContainerInitializer;
-import java.lang.System.Logger;
-import java.lang.System.Logger.Level;
-import static java.lang.System.Logger.Level.DEBUG;
-import java.lang.reflect.InvocationTargetException;
-
-/**
- * The WebApplicationExtension that adds the StandardTempDirInitializer.
- *
- * @author Manfred Riem (mriem@manorrock.com)
- */
-public class StandardTempDirExtension implements WebApplicationExtension {
-
-    /**
-     * Stores the logger.
-     */
-    private static final Logger LOGGER = System.getLogger(StandardTempDirExtension.class.getName());
-
-    @Override
-    public void configure(WebApplication webApplication) {
-        LOGGER.log(DEBUG, "Adding the StandardTempDirInitializer");
-        try {
-            ClassLoader classLoader = webApplication.getClassLoader();
-
-            Class<? extends ServletContainerInitializer> clazz
-                    = classLoader.
-                            loadClass(StandardTempDirInitializer.class.getName())
-                            .asSubclass(ServletContainerInitializer.class);
-
-            ServletContainerInitializer initializer
-                    = clazz.getDeclaredConstructor().newInstance();
-
-            webApplication.addInitializer(initializer);
-
-        } catch (ClassNotFoundException | NoSuchMethodException | SecurityException
-                | InstantiationException | IllegalAccessException
-                | IllegalArgumentException | InvocationTargetException ex) {
-            LOGGER.log(Level.WARNING, "Unable to add the StandardTempDirInitializer", ex);
-        }
-    }
-}
